+++ src/components/index.ts
@@ -13,9 +13,6 @@ export * from './setup-card';
 export { default as SetupCardHeader } from './setup-card-header';
 export * from './setup-card-header';
 
-export { default as SetupCardChip } from './setup-card-chip';
-export * from './setup-card-chip';
-
 export { default as SetupCardTask } from './setup-card-task';
 export * from './setup-card-task';
 
@@ -45,6 +42,9 @@ export * from './form-control-label';
 
 export { default as SetupCardListItem } from './setup-card-list-item';
 export * from './setup-card-list-item';
+
+export { default as Chip } from './chip';
+export * from './chip';
 
 export { default as ProgressBar } from './progress-bar';
 export * from './progress-bar';
