--- src/components/setup-card-chip/index.ts
@@ -1,1 +1,0 @@
-export { default } from './setup-card-chip';
