+++ src/components/setup-card-header/setup-card-header.stories.tsx
@@ -1,13 +1,10 @@
 import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
+import { SetupCardHeader, Chip } from '..';
+import * as ChipStories from '../chip/chip.stories';
 
-import { SetupCardHeader, SetupCardChip } from '..';
-
-import * as SetupCardChipStories from '../setup-card-chip/setup-card-chip.stories';
-
-// More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-  title: 'Components/SetupCardHeader',
+  title: 'Setup Cards/SetupCardHeader',
   component: SetupCardHeader,
   decorators: [(story) => <div style={{ maxWidth: 800, margin: '0 auto' }}>{story()}</div>],
   parameters: {
@@ -36,10 +33,9 @@ const Template: ComponentStory<typeof SetupCardHeader> = (args) => (
   <SetupCardHeader {...args} />
 );
 
-// More on args: https://storybook.js.org/docs/react/writing-stories/args
 export const Header = Template.bind({});
 Header.args = {
   title: 'Example Card Header',
   subheader: 'Donec ullamcorper nulla non metus auctor fringilla.',
-  action: <SetupCardChip {...SetupCardChipStories.Todo.args} />,
+  action: <Chip {...ChipStories.Todo.args} />,
 };
