+++ src/components/setup-card-list-item/setup-card-list-item.stories.tsx
@@ -1,14 +1,11 @@
 import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 import DownhillSkiingIcon from '@mui/icons-material/DownhillSkiing';
-
 import { SetupCardList, SetupCardListItem } from '..';
 
-// More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-  title: 'Components/SetupCardListItem',
+  title: 'Setup Cards/SetupCardListItem',
   component: SetupCardListItem,
-  // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {
     icon: {
       control: 'boolean',
@@ -19,7 +16,6 @@ export default {
   },
 } as ComponentMeta<typeof SetupCardListItem>;
 
-// More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
 const Template: ComponentStory<typeof SetupCardListItem> = (args) => (
   <SetupCardList>
     <SetupCardListItem {...args} />
