+++ src/components/setup-card/setup-card.stories.mdx
@@ -29,7 +29,7 @@ import * as SetupCardFooterStories from '../setup-card-footer/setup-card-footer.
 
 import { theme } from '../../theme';
 
-<Meta title="Components/SetupCard" component={SetupCard} />
+<Meta title="Setup Cards/SetupCard" component={SetupCard} />
 
 # SetupCard
 Various setup tasks are set within cards on the WME Set Up page.
